Add River OpenTelemetry tracing implementation
Here, add OpenTelemetry support for River with a middleware that will do
instrumentation on job insertion and job working.

The impetus is that we've had multiple requests for DataDog support now.
I wrote an equivalent middleware using DataDog's Go APIs, but found them
to be fairly unwieldy, with lots and lots of dependencies [1], and
DataDog appears to have first class OpenTelemetry support anyway.

There's also some potential for more sophisticated functionality here
like being able to trace a job in a distributed sense from when it's
inserted to when it's worked. I've opted to punt on that for now so we
can get a basic version of this together for testing sooner rather than
later, and under the principle of not developing too many features
prospectively that we don't know whether people will actually find
useful or not (let's see what people say about it).

[1] https://x.com/brandur/status/1898147112837357762