`insert_many` metric + further name refinement
Drop the `job_` prefix from all metrics given they're redundant and the
names of spans doesn't have the same. Rename insert metrics to
`insert_many*` where appropriate and add an `insert_many` counter that
differs fro the `insert` counter in that it counts insertion of batches
rather than insertion of individual jobs.

Overall, this should be more consistent, and not get us boxed in any
particular naming discrepancies relating to an individual job versus a
batch.