Add option to make metric duration unit configurable + testing (#10)
Here, bring in a new option `MiddlewareConfig.DurationUnit` which can be
used to change the default duration of duration-based metrics from
seconds to milliseconds. There seems to be a reasonable number of people
in both the seconds and milliseconds camps [1], so it seems okay to make
this configurable.

The first release also didn't bother testing metrics (since at the time,
I couldn't find a testing package for metrics like existed for traces),
so we also bring in metrics tests for the existing behavior along with
the new `DurationUnit` option.

Also fixes a small bug wherein attributes weren't being assigned as
expected to a couple of the metrics.

Fixes #9.

[1] https://github.com/open-telemetry/opentelemetry-specification/issues/2977