Don't emit higher cardinality attributes for metrics (#14)
This ones follows up #11. I said in the pull request that higher
cardinality attributes aren't emitted with metrics, but I got that wrong
as I'd forgotten I was feeding all span attributes into metrics as well.

Here, bifurcate attributes so that high cardinality attributes are added
to spans, but not to metrics.