Add `EnableSemanticMetrics` setting to emit "semantic" metrics (#13)
Here, address #7 to bring in a new option `EnableSemanticMetrics` that
causes the middleware to emit metrics compliant with OpenTelemetry's
"semantic conventions" [1]. These are a way to have all pub/sub systems
emit metrics with the same name and differentiated by attribute.

I've put this behind a configuration option because I'm not super
convinced that this is going to be an important feature for most people,
and under the assumption that metrics are cheap but not free, it
probably makes sense to emit only as many as necessary.

Fixes #7.

[1] https://opentelemetry.io/docs/specs/semconv/messaging/messaging-metrics/