Pure DataDog version of telemetry middleware
Just putting this one up here for posterity because I'm intending to
delete it locally. This is a version of the middleware from before
OpenTelemetry that uses DataDog's SDK directly. It probably would've
been okay, but we went with OpenTelemetry in the end because it's fully
compatible with DataDog, gives us broader ecosystem access, and requires
fewer dependencies.