Add `nilerror` hook to detect nil structs wrapped in non-nil error interfaces (#25)
This one's aimed at helping people debug problems like the one raised
here [1]. There's a common mistake in Go where a nil error-compliant
struct is wrapped in a non-nil error interface and unexpectedly caught
by a `if err != nil { ... }` check. For example:

    type CustomError {
        InternalErr error
    }

    type JobWorker struct {
        river.WorkerDefaults[workers.JobWorkerArgs
    }

    func (jw *JobWorker) Work(ctx context.Context, job *river.Job[workers.JobWorkerArgs]) error {
        var customErr *CustomError
        return customErr
    }

This is such a common problem in Go that the language maintains a long
FAQ on the topic: https://go.dev/doc/faq#nil_error

We probably don't want to bake a solution for this right into core River
because there'd be huge room for disagreement for what to do about it.
Purists would not want the case handled because technically Go is
behaving the way that it's supposed to. Beyond that, in case a problem
is found, it's debatable whether an error should be emitted (so it can
be easily caught in tests) or just logged.

Here, we add a hook in `rivercontrib` as a nice compromise. People who
want extra checks on this problem can easily configure it, and everyone
else can ignore it.

This requires a new `HookWorkEnd` interface in core River to work [2].
It could've been added as a middleware instead, but it seems to me that
a lighter check that doesn't go on the stack in a hook is the better
approach.

[1] https://github.com/riverqueue/river/issues/806
[2] https://github.com/riverqueue/river/pull/863