Fix GitHub import path for `nilerror` package (#29)
I just realizing that the import path for `nilerror` in its `go.mod` is
completely wrong due to the omission of `riverqueue` from the URL,
meaning that external packages won't be able to import it.

I guess that means no one is really using it, but still, here fix up the
problem.