Add `versionedjob` contrib package with basic versioning framework
Here, add a contrib package called `versionedjob` that brings in a very
basic job versioning framework. The framework is very lightweight and
users could easily have written it themselves, but it might be a little
non-obvious how to leverage hooks to bring it in.

Contrib should be a really nice home for this feature because it's a
little too trivial to bother with in core, but it's nice for people to
have a semi-official reference that's well-maintained.